// projects/sdp-api/src/lib/partner/models/contract-info-detail.ts
/* tslint:disable */
export interface ContractInfoDetail {
  accDigitalAssets?: number;
  accRating?: number;
  assetRatingPer?: number;
  atxDigitalAssets?: number;
  atxRating?: number;
  contractNumber?: string;
  contractStatus?: string;
  contractValue?: number;
  customerId?: string;
  customerName?: string;
  endDate?: string;
  partnerId?: string;
  portalAccess?: boolean;
  region?: string;
  solution?: string;
  startDate?: string;
  totalDigitalAssets?: number;
  useCase?: string;
}